break apart budget-category-features
