--- src/extension/features/budget/budget-category-features/goal-indicator/settings.js
@@ -1,9 +1,0 @@
-module.exports = {
-  name: 'GoalIndicator',
-  type: 'checkbox',
-  default: false,
-  section: 'budget',
-  title: 'Add Goals Indication',
-  description:
-    'Add indicators for subcategories with goals. Types: (M)onthly goal,  target by (D)ate goal, (T)arget without date, and (U)pcoming transactions.',
-};
