--- src/extension/features/budget/budget-category-features/goal-warning-color/settings.js
@@ -1,9 +1,0 @@
-module.exports = {
-  name: 'GoalWarningColor',
-  type: 'checkbox',
-  default: false,
-  section: 'budget',
-  title: 'Goal Indicator Warning Color',
-  description:
-    'Change the orange goal underfunded warning to blue, to better differentiate it from credit card overspending.',
-};
