--- src/extension/features/budget/budget-category-features/highlight-negatives/index.css
@@ -1,36 +1,0 @@
-/* row */
-.budget-table-row[data-toolkit-negative-available]
-  .budget-table-cell-available
-  .ynab-new-budget-available-number {
-  background: var(--budget_balance_negative_background) !important;
-  color: var(--budget_balance_negative_text) !important;
-}
-.budget-table-row[data-toolkit-negative-available]
-  .budget-table-cell-available
-  .ynab-new-budget-available-number:hover {
-  background: var(--budget_balance_negative_background_active) !important;
-}
-
-/* pacing column */
-li.budget-table-cell-pacing span.cautious {
-  background: var(--budget_balance_negative_background) !important;
-  color: var(--budget_balance_negative_text) !important;
-}
-li.budget-table-cell-pacing .budget-table-cell-pacing-display.cautious:hover {
-  background: var(--budget_balance_negative_background_active) !important;
-}
-li.budget-table-cell-pacing span.deemphasized.cautious,
-li.budget-table-cell-pacing span.deemphasized.cautious:hover {
-  background: none !important;
-}
-
-/* inspector */
-.budget-inspector[data-toolkit-negative-available] .budget-inspector-category-overview dt.cautious {
-  color: var(--budget_balance_negative_text);
-}
-.budget-inspector[data-toolkit-negative-available]
-  .budget-inspector-category-overview
-  dd
-  .cautious {
-  background: var(--budget_balance_negative_background);
-}
