--- src/extension/features/budget/budget-category-features/settings.js
@@ -1,9 +1,0 @@
-module.exports = {
-  name: 'BudgetCategoryFeatures',
-  section: 'system',
-  default: true,
-  type: 'system',
-  title: 'BudgetCategoryFeatures',
-  description:
-    'This is a system feature which automatically adds data to budget-category-rows which can be used by other Toolkit features. System features do not appear in the options menu.',
-};
