--- src/extension/features/budget/budget-category-features/target-balance-warning/settings.js
@@ -1,9 +1,0 @@
-module.exports = {
-  name: 'TargetBalanceWarning',
-  type: 'checkbox',
-  default: false,
-  section: 'budget',
-  title: 'Warn When Target Balance is Not Reached',
-  description:
-    'Will highlight balances of categories with Target Balances that have not yet been met.',
-};
