+++ src/extension/features/budget/goal-indicator/index.css
@@ -1,8 +1,8 @@
-.toolkit-goal-indicator {
+.tk-goal-indicator {
   color: gray;
   padding-left: 0.5em;
 }
 
-.toolkit-goal-indicator:last-child {
+.tk-goal-indicator:last-child {
   padding-right: 0.5em;
 }
