+++ src/extension/features/budget/goal-indicator/settings.js
@@ -1,0 +1,9 @@
+module.exports = {
+  name: 'GoalIndicator',
+  type: 'checkbox',
+  default: false,
+  section: 'budget',
+  title: 'Add Goals Indication',
+  description:
+    'Add an indicator to subcategories with goals. Types: (M)onthly goal, target by (D)ate goal, (T)arget without date, (S)pending goals, and (U)pcoming transactions.',
+};
