+++ src/extension/features/budget/goal-warning-color/settings.js
@@ -1,0 +1,9 @@
+module.exports = {
+  name: 'GoalWarningColor',
+  type: 'checkbox',
+  default: false,
+  section: 'budget',
+  title: 'Distinguish Underfunded Goals',
+  description:
+    'Change the default orange goal underfunded warning to blue to better differentiate it from credit card overspending.',
+};
