+++ src/extension/features/budget/highlight-negatives/index.css
@@ -1,0 +1,10 @@
+.budget-table-row .ynab-new-budget-available-number.cautious.credit {
+  background-color: var(--budget_balance_negative_background) !important;
+  color: var(--budget_balance_negative_text) !important;
+}
+
+body[data--square-negative-mode]
+  .budget-table-row
+  .ynab-new-budget-available-number.cautious.credit {
+  border-radius: 0rem;
+}
