+++ src/extension/features/budget/target-balance-warning/settings.js
@@ -1,0 +1,9 @@
+module.exports = {
+  name: 'TargetBalanceWarning',
+  type: 'checkbox',
+  default: false,
+  section: 'budget',
+  title: 'Emphasize Unmet Target Balances',
+  description:
+    'Add warning emphasis to categories with Target Balances that have not yet been met.',
+};
